Break utils file
